requirements.txt
aniso8601==9.0.1
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
flask-testing==0.8.1
googletrans==3.0.0
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
helpers==0.2.0
hpack==3.0.0
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
mypy==1.2.0
mypy-extensions==1.0.0
packaging==23.0
pluggy==1.0.0
PyJWT==2.6.0
PyQt6==6.4.1
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
pyrsistent==0.19.3
pytest==7.4.0
pytest-mock==3.11.1
pytz==2022.7.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.4
telebot==0.0.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.12
Werkzeug==2.2.3
wikipedia==1.4.0
zipp==3.14.0
